Reorganize test structure into logical subdirectories
- Create Core/, Generation/, and Serialization/ subdirectories
- Move test files to appropriate directories by functional area:
  * Core/: RangeFinderTests, RangeFinderFactoryTests, RangeTreeCompatibilityTests
  * Generation/: Dataset generation, analysis, and parameter validation tests
  * Serialization/: CSV, Parquet, and file loading tests
- Update namespaces to match new directory structure
- Add proper using statements for TestBase inheritance
- Maintain TestBase.cs at root for easy access across all namespaces

All 165 tests continue to pass. Improves maintainability and developer navigation.
