+++ RangeFinder.Tests/Generation/QueryGenerationTests.cs
@@ -1,8 +1,9 @@
 using RangeFinder.Core;
 using RangeFinder.IO.Generation;
+using RangeFinder.Tests;
 using System.Reflection;
 
-namespace RangeFinder.Tests;
+namespace RangeFinder.Tests.Generation;
 
 /// <summary>
 /// Tests for query generation functionality.
