+++ RangeFinder.Tests/Serialization/RangeSerializerCsvTests.cs
@@ -2,15 +2,15 @@ using RangeFinder.Core;
 using RangeFinder.IO;
 using RangeFinder.IO.Serialization;
 
-namespace RangeFinder.Tests;
+namespace RangeFinder.Tests.Serialization;
 
 [TestFixture]
-public class RangeSerializerParquetTests
+public class RangeSerializerCsvTests
 {
-    private string GetTempFilePath() => Path.GetTempFileName().Replace(".tmp", ".parquet");
+    private string GetTempFilePath() => Path.GetTempFileName().Replace(".tmp", ".csv");
 
     [Test]
-    public void ParquetSaveAndLoad_IntegerRanges_PreservesData()
+    public void CsvSaveAndLoad_IntegerRanges_PreservesData()
     {
         var tempFilePath = GetTempFilePath();
         try
@@ -22,8 +22,8 @@ public class RangeSerializerParquetTests
                 new NumericRange<int, string>(40, 50, "Range3")
             };
 
-            originalRanges.WriteParquet(tempFilePath);
-            var loadedRanges = RangeSerializer.ReadParquet<int, string>(tempFilePath).ToList();
+            originalRanges.WriteCsv(tempFilePath);
+            var loadedRanges = RangeSerializer.ReadCsv<int, string>(tempFilePath).ToList();
 
             Assert.That(loadedRanges, Is.EqualTo(originalRanges));
         }
@@ -35,7 +35,7 @@ public class RangeSerializerParquetTests
     }
 
     [Test]
-    public void ParquetSaveAndLoad_DoubleRanges_PreservesData()
+    public void CsvSaveAndLoad_DoubleRanges_PreservesData()
     {
         var tempFilePath = GetTempFilePath();
         try
@@ -47,10 +47,17 @@ public class RangeSerializerParquetTests
                 new NumericRange<double, int>(40.2, 50.8, 300)
             };
 
-            originalRanges.WriteParquet(tempFilePath);
-            var loadedRanges = RangeSerializer.ReadParquet<double, int>(tempFilePath).ToList();
+            originalRanges.WriteCsv(tempFilePath);
+            var loadedRanges = RangeSerializer.ReadCsv<double, int>(tempFilePath).ToList();
 
-            Assert.That(loadedRanges, Is.EqualTo(originalRanges));
+            // For floating point, we need tolerance comparison
+            Assert.Multiple(() =>
+            {
+                Assert.That(loadedRanges, Has.Count.EqualTo(3));
+                Assert.That(loadedRanges.Select(r => r.Start), Is.EqualTo(originalRanges.Select(r => r.Start)).Within(0.001));
+                Assert.That(loadedRanges.Select(r => r.End), Is.EqualTo(originalRanges.Select(r => r.End)).Within(0.001));
+                Assert.That(loadedRanges.Select(r => r.Value), Is.EqualTo(originalRanges.Select(r => r.Value)));
+            });
         }
         finally
         {
@@ -60,7 +67,7 @@ public class RangeSerializerParquetTests
     }
 
     [Test]
-    public void ParquetSaveAndLoad_StringWithCommas_HandlesEscaping()
+    public void CsvSaveAndLoad_StringWithCommas_HandlesEscaping()
     {
         var tempFilePath = GetTempFilePath();
         try
@@ -72,8 +79,8 @@ public class RangeSerializerParquetTests
                 new NumericRange<int, string>(40, 50, "Normal value")
             };
 
-            originalRanges.WriteParquet(tempFilePath);
-            var loadedRanges = RangeSerializer.ReadParquet<int, string>(tempFilePath).ToList();
+            originalRanges.WriteCsv(tempFilePath);
+            var loadedRanges = RangeSerializer.ReadCsv<int, string>(tempFilePath).ToList();
 
             Assert.That(loadedRanges, Is.EqualTo(originalRanges));
         }
@@ -85,7 +92,7 @@ public class RangeSerializerParquetTests
     }
 
     [Test]
-    public async Task ParquetSaveAndLoadAsync_IntegerRanges_PreservesData()
+    public async Task CsvSaveAndLoadAsync_IntegerRanges_PreservesData()
     {
         var tempFilePath = GetTempFilePath();
         try
@@ -97,8 +104,8 @@ public class RangeSerializerParquetTests
                 new NumericRange<int, string>(40, 50, "Range3")
             };
 
-            await originalRanges.WriteParquetAsync(tempFilePath);
-            var loadedRanges = (await RangeSerializer.ReadParquetAsync<int, string>(tempFilePath)).ToList();
+            await originalRanges.WriteCsvAsync(tempFilePath);
+            var loadedRanges = (await RangeSerializer.ReadCsvAsync<int, string>(tempFilePath)).ToList();
 
             Assert.That(loadedRanges, Is.EqualTo(originalRanges));
         }
@@ -110,15 +117,15 @@ public class RangeSerializerParquetTests
     }
 
     [Test]
-    public void ParquetSaveAndLoad_EmptyCollection_CreatesEmptyFile()
+    public void CsvSaveAndLoad_EmptyCollection_CreatesEmptyFile()
     {
         var tempFilePath = GetTempFilePath();
         try
         {
             var originalRanges = Array.Empty<NumericRange<int, string>>();
 
-            originalRanges.WriteParquet(tempFilePath);
-            var loadedRanges = RangeSerializer.ReadParquet<int, string>(tempFilePath).ToList();
+            originalRanges.WriteCsv(tempFilePath);
+            var loadedRanges = RangeSerializer.ReadCsv<int, string>(tempFilePath).ToList();
 
             Assert.That(loadedRanges, Is.Empty);
         }
@@ -130,7 +137,7 @@ public class RangeSerializerParquetTests
     }
 
     [Test]
-    public void ParquetSaveAndLoad_NullValues_HandlesCorrectly()
+    public void CsvSaveAndLoad_NullValues_HandlesCorrectly()
     {
         var tempFilePath = GetTempFilePath();
         try
@@ -142,10 +149,17 @@ public class RangeSerializerParquetTests
                 new NumericRange<int, string?>(40, 50, null)
             };
 
-            originalRanges.WriteParquet(tempFilePath);
-            var loadedRanges = RangeSerializer.ReadParquet<int, string?>(tempFilePath).ToList();
+            originalRanges.WriteCsv(tempFilePath);
+            var loadedRanges = RangeSerializer.ReadCsv<int, string?>(tempFilePath).ToList();
 
-            Assert.That(loadedRanges, Is.EqualTo(originalRanges));
+            // CsvHelper converts null to empty string
+            var expectedRanges = new[]
+            {
+                new NumericRange<int, string?>(1, 10, string.Empty),
+                new NumericRange<int, string?>(20, 30, "Range2"),
+                new NumericRange<int, string?>(40, 50, string.Empty)
+            };
+            Assert.That(loadedRanges, Is.EqualTo(expectedRanges));
         }
         finally
         {
